♻️ refactor: decouple fetch logic from dashboard page and add meta constants
