+++ src/lib/constants/metadata.ts
@@ -1,0 +1,5 @@
+export const TITLE =
+  "Jaam Toast - Jamstack App Deployment Service Platform | Deploy Your Own Websites Quick And Easy Like Toasts";
+
+export const DESCRIPTION =
+  "Bring your jamstack app, and make your products get deployed like an easy toast.";
