+++ src/lib/hooks/useFetchDeployment.ts
@@ -1,0 +1,45 @@
+import { useEffect } from "react";
+import { useRecoilState } from "recoil";
+import { setCookie } from "cookies-next";
+
+import { getUserDeployments } from "./../api/index";
+import userDeploymentsState from "../recoil/userDeployments";
+
+import { UserDeploymentData } from "./../../types/index";
+
+function useFetchDeployment(userId: string) {
+  const [userDeploymentsList, setUserDeploymentsList] =
+    useRecoilState<UserDeploymentData[]>(userDeploymentsState);
+
+  useEffect(() => {
+    if (!userId) return;
+
+    const handleUserDeployments = async () => {
+      try {
+        const { data: userDeployments } = await getUserDeployments(userId);
+
+        const copyUserDeployments: UserDeploymentData[] = JSON.parse(
+          JSON.stringify(userDeployments),
+        );
+
+        const filteredUserDeployments = copyUserDeployments.map(deployData => {
+          const filteredDeployData = deployData;
+          filteredDeployData.buildingLog = [];
+
+          return filteredDeployData;
+        });
+
+        setUserDeploymentsList(userDeployments);
+        setCookie("userDeployments", JSON.stringify(filteredUserDeployments));
+      } catch (error) {
+        console.info(error);
+      }
+    };
+
+    handleUserDeployments();
+  }, [setUserDeploymentsList, userId]);
+
+  return userDeploymentsList;
+}
+
+export default useFetchDeployment;
